+++ src/Components/Mint/index.tsx
@@ -1,4 +1,4 @@
-import { nearWallet} from "../wallet/wallet";
+import { nearWallet} from "../../libs/wallet";
 import {randomSkin} from "../../config/skins"
 import {Contract, utils, providers} from "near-api-js";
 import { Button } from "../../styles/styles";
