+++ src/Components/Nft/styles.ts
