+++ src/libs/wallet.tsx
@@ -1,4 +1,4 @@
-import getConfig from "../../config/config";
+import getConfig from "../config/config";
 import * as nearAPI from "near-api-js"
 
 const CONTRACT_NAME = process.env.CONTRACT_NAME || "dev-1644775935237-43621084292867"; // "katesonia2.testnet";
